refactor: Move port mapping code to its own module
This commit does not change behavior.
